Develop

+++ EssentialConnection/EssentialConnection/Controllers/AlunosController.cs
@@ -1,5 +1,4 @@
-﻿#nullable disable
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -7,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using EssentialConnection.Models;
+using Microsoft.AspNet.Identity;
 
 namespace EssentialConnection.Controllers
 {
@@ -29,7 +29,7 @@ namespace EssentialConnection.Controllers
         // GET: Alunos/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Aluno == null)
             {
                 return NotFound();
             }
@@ -57,11 +57,20 @@ namespace EssentialConnection.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("AlunoID,Nome,Endereco,Email,DataNascimento,Telefone,Matricula,CursoId,CurriculoId,Login,Senha")] Aluno aluno)
+        public async Task<IActionResult> Create(string nome,string email, string telefone, int cursoId, string userId)
         {
+            Aluno aluno = new Aluno();
+            aluno.NomeCompleto = nome;
+            aluno.email = email;
+            aluno.Telefone = telefone;
+            aluno.CursoId = cursoId;
+            aluno.UserId = userId;
+            if (ModelState.IsValid)
+            {
                 _context.Add(aluno);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
+            }
             ViewData["CursoId"] = new SelectList(_context.Curso, "CursoID", "CursoID", aluno.CursoId);
             return View(aluno);
         }
@@ -69,7 +78,7 @@ namespace EssentialConnection.Controllers
         // GET: Alunos/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Aluno == null)
             {
                 return NotFound();
             }
@@ -88,13 +97,15 @@ namespace EssentialConnection.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("AlunoID,Nome,Endereco,Email,DataNascimento,Telefone,Matricula,CursoId,CurriculoId,Login,Senha")] Aluno aluno)
+        public async Task<IActionResult> Edit(int id, [Bind("AlunoID,NomeCompleto,UserId,Telefone,CursoId,CurriculoId")] Aluno aluno)
         {
             if (id != aluno.AlunoID)
             {
                 return NotFound();
             }
 
+            if (ModelState.IsValid)
+            {
                 try
                 {
                     _context.Update(aluno);
@@ -111,6 +122,8 @@ namespace EssentialConnection.Controllers
                         throw;
                     }
                 }
+                return RedirectToAction(nameof(Index));
+            }
             ViewData["CursoId"] = new SelectList(_context.Curso, "CursoID", "CursoID", aluno.CursoId);
             return View(aluno);
         }
@@ -118,7 +131,7 @@ namespace EssentialConnection.Controllers
         // GET: Alunos/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Aluno == null)
             {
                 return NotFound();
             }
@@ -139,15 +152,31 @@ namespace EssentialConnection.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Aluno == null)
+            {
+                return Problem("Entity set 'Context.Aluno'  is null.");
+            }
             var aluno = await _context.Aluno.FindAsync(id);
-            _context.Aluno.Remove(aluno);
+            if (aluno != null)
+            {
+                _context.Aluno.Remove(aluno);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool AlunoExists(int id)
         {
-            return _context.Aluno.Any(e => e.AlunoID == id);
+          return (_context.Aluno?.Any(e => e.AlunoID == id)).GetValueOrDefault();
+        }
+
+        public void AdicionaCurriculo(int curriculo, int alunoId)
+        {
+            var aluno = _context.Aluno.FirstOrDefault(x => x.AlunoID == alunoId);
+            aluno.CurriculoId = curriculo;
+            _context.Update(aluno);
+            _context.SaveChangesAsync();
         }
     }
 }
